[30-bag-fix-with-auth] Moved balance to a separate Slice
