+++ client/src/hooks/useHeaderLogic.ts
@@ -2,7 +2,8 @@ import {useEffect, useState} from 'react';
 import {useAppDispatch, useAppSelector} from "../store/hooks.ts";
 import {openModal} from "../store/modal/modalSlice.ts";
 import type {RootState} from "../store/store.ts";
-import {getBalanceByUser, logout} from "../store/user/userSlice.ts";
+import {getBalanceByUser} from "../store/balance/balanceSlice.ts";
+import {logout} from "../store/user/userSlice.ts";
 import type {Balance} from "../types";
 
 export type UseHeaderLogic = {
@@ -17,7 +18,7 @@ export type UseHeaderLogic = {
 
 export const useHeaderLogic = (): UseHeaderLogic => {
     const auth = useAppSelector((state: RootState) => state.user.isAuthenticated);
-    const balance = useAppSelector((state: RootState) => state.user.balance);
+    const balance = useAppSelector((state: RootState) => state.balance);
     const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
     const dispatch = useAppDispatch();
